requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./bin/compile_dependencies.sh
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via glom
billiard==3.6.4.0
    # via celery
boltons==23.0.0
    # via
    #   face
    #   glom
celery==5.2.7
    # via
    #   -r requirements/base.in
    #   notifications-api-common
certifi==2022.12.7
    # via
    #   django-simple-certmanager
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via
    #   django-loose-fk
    #   drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==39.0.2
    # via
    #   django-simple-certmanager
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
django==3.2.18
    # via
    #   -r requirements/base.in
    #   django-admin-index
    #   django-axes
    #   django-choices
    #   django-filter
    #   django-loose-fk
    #   django-markup
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rest-framework-condition
    #   django-rosetta
    #   django-sendfile2
    #   django-simple-certmanager
    #   django-sniplates
    #   django-solo
    #   djangorestframework
    #   drf-nested-routers
    #   drf-yasg
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
django-admin-index==2.0.2
    # via -r requirements/base.in
django-axes==5.40.1
    # via -r requirements/base.in
django-better-admin-arrayfield==1.4.2
    # via
    #   -r requirements/base.in
    #   mozilla-django-oidc-db
django-choices==1.7.2
    # via
    #   -r requirements/base.in
    #   vng-api-common
django-cors-middleware==1.5.0
    # via -r requirements/base.in
django-filter==2.4.0
    # via
    #   -r requirements/base.in
    #   django-loose-fk
    #   vng-api-common
django-ipware==5.0.0
    # via django-axes
django-loose-fk==1.0.3
    # via -r requirements/base.in
django-markup==1.3
    # via vng-api-common
django-ordered-model==3.7.4
    # via django-admin-index
django-privates==1.5.0
    # via django-simple-certmanager
django-redis==5.2.0
    # via -r requirements/base.in
django-relativedelta==1.1.2
    # via zgw-consumers
django-rest-framework-condition==0.1.1
    # via vng-api-common
django-rosetta==0.9.8
    # via -r requirements/base.in
django-sendfile2==0.7.0
    # via django-privates
django-simple-certmanager==1.3.0
    # via zgw-consumers
django-sniplates==0.7.0
    # via -r requirements/base.in
django-solo==2.0.0
    # via
    #   mozilla-django-oidc-db
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   django-loose-fk
    #   drf-nested-routers
    #   drf-yasg
    #   notifications-api-common
    #   vng-api-common
djangorestframework-camel-case==1.4.2
    # via
    #   notifications-api-common
    #   vng-api-common
drf-nested-routers==0.93.4
    # via vng-api-common
drf-yasg==1.21.5
    # via vng-api-common
elastic-apm==6.15.1
    # via -r requirements/base.in
face==20.1.1
    # via glom
faker==18.3.0
    # via zgw-consumers
gemma-zds-client==1.0.1
    # via
    #   -r requirements/base.in
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
glom==23.1.1
    # via mozilla-django-oidc-db
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
iso-639==0.4.5
    # via vng-api-common
isodate==0.6.1
    # via vng-api-common
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
josepy==1.13.0
    # via mozilla-django-oidc
kombu==5.2.4
    # via celery
markdown==3.4.3
    # via vng-api-common
markupsafe==2.1.2
    # via jinja2
mozilla-django-oidc==2.0.0
    # via mozilla-django-oidc-db
mozilla-django-oidc-db==0.12.0
    # via -r requirements/base.in
notifications-api-common==0.2.1
    # via vng-api-common
oyaml==1.0
    # via vng-api-common
packaging==23.0
    # via drf-yasg
polib==1.2.0
    # via django-rosetta
prompt-toolkit==3.0.38
    # via click-repl
psycopg2==2.9.5
    # via -r requirements/base.in
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via
    #   gemma-zds-client
    #   vng-api-common
pyopenssl==23.0.0
    # via
    #   django-simple-certmanager
    #   josepy
    #   zgw-consumers
python-dateutil==2.8.2
    # via
    #   django-relativedelta
    #   faker
python-decouple==3.8
    # via -r requirements/base.in
python-dotenv==1.0.0
    # via -r requirements/base.in
pytz==2022.7.1
    # via
    #   -r requirements/base.in
    #   celery
    #   django
    #   drf-yasg
pyyaml==6.0
    # via
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
redis==4.5.3
    # via django-redis
requests==2.28.2
    # via
    #   coreapi
    #   django-rosetta
    #   gemma-zds-client
    #   mozilla-django-oidc
    #   requests-mock
    #   vng-api-common
    #   zgw-consumers
requests-mock==1.10.0
    # via zgw-consumers
ruamel-yaml==0.17.21
    # via drf-yasg
sentry-sdk==1.17.0
    # via -r requirements/base.in
six==1.16.0
    # via
    #   click-repl
    #   django-choices
    #   django-markup
    #   isodate
    #   python-dateutil
    #   requests-mock
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.15
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.21
    # via -r requirements/base.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
vng-api-common[markdown_docs]==1.9.0
    # via -r requirements/base.in
wcwidth==0.2.6
    # via prompt-toolkit
wrapt==1.15.0
    # via elastic-apm
zgw-consumers==0.25.0
    # via
    #   -r requirements/base.in
    #   notifications-api-common

# The following packages are considered to be unsafe in a requirements file:
setuptools==69.0.2
    # via
    #   django-axes
    #   josepy
